requirements.txt
chromadb
pandas
numpy
scipy
statsmodels
matplotlib
seaborn
google-generativeai
python-dotenv
flask
streamlit